interface check fix
in order to have relible interface check we expect 0x0E value returned
back from init function. But if the radio was previously initialized
and received some data, this can result as returning 0x40 which harder
to check. Now flushing rx buffers were added to be able check only for
0x0E value. Other values meaning an interface error or chip malfunction.

Signed-off-by: Pavel <pasha-net@narod.ru>